1. renamed HashCodeFinderTest to HashCodeMethodFinderTest to match the follow the general junit convention [classname]Test 2. modified EqualsMethodTextCreator to move the call to super.equals(...) check from a separate if branch to the return statement. From: {code} if (!super.equals(obj)) { return false; } {code} To: {code} return super.equals(obj) [&& [EQUALS_METHOD](field1, other.field1) ...] {code} 3. updated appropriate assertions in EqualsMethodTextCreatorTest
